pyproject.toml
[tool.poetry]
name = "upvote-scripts"
version = "0.1.0"
description = ""
authors = ["Gin Fuyou <42@doratoa.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
dateutils = "^0.6.12"
ruff = "^0.1.3"


[tool.ruff]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "W", "F", "I001", "D"]
ignore = ["D103", "D212"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["F4", "W", "I001", "E7", "D"]
unfixable = []

line-length = 128

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
